Adjust namespace.
