+++ osu.Game.Rulesets.Karaoke.Tests/Edit/TestSceneLyricEditorScreen.cs
@@ -10,12 +10,12 @@ using osu.Game.Overlays;
 using osu.Game.Rulesets.Edit;
 using osu.Game.Rulesets.Karaoke.Beatmaps.Formats;
 using osu.Game.Rulesets.Karaoke.Edit.ImportLyric;
-using osu.Game.Rulesets.Karaoke.Edit.LyricEditor;
 using osu.Game.Rulesets.Karaoke.Tests.Beatmaps;
 using osu.Game.Rulesets.Karaoke.Tests.Resources;
 using osu.Game.Screens.Edit;
 using osu.Game.Tests.Visual;
 using System.IO;
+using osu.Game.Rulesets.Karaoke.Edit.Lyric;
 
 namespace osu.Game.Rulesets.Karaoke.Tests.Edit
 {
