+++ osu.Game.Rulesets.Karaoke/Edit/Lyric/Components/Badges/LayoutInfoBadge.cs
@@ -3,26 +3,25 @@
 
 using osu.Framework.Allocation;
 using osu.Game.Graphics;
-using osu.Game.Rulesets.Karaoke.Objects;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.LyricEditor.Components.Badges
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyric.Components.Badges
 {
-    public class StyleInfoBadge : Badge
+    public class LayoutInfoBadge : Badge
     {
-        public StyleInfoBadge(Lyric lyric)
+        public LayoutInfoBadge(Objects.Lyric lyric)
             : base(lyric)
         {
-            lyric.SingersBindable.BindValueChanged(value =>
+            lyric.LayoutIndexBindable.BindValueChanged(value =>
             {
-                var newStyleIndex = value.NewValue;
-                BadgeText = $"Singer : {newStyleIndex}";
+                var newLayoutIndex = value.NewValue;
+                BadgeText = $"Layout : {newLayoutIndex}";
             }, true);
         }
 
         [BackgroundDependencyLoader]
         private void load(OsuColour colours)
         {
-            BadgeColour = colours.Green;
+            BadgeColour = colours.Pink;
         }
     }
 }
